collect commen functions in utils for visualisations
